requirements.txt
tensorflow==2.16.2
keras==3.10.0
numpy==1.26.4
pillow==11.3.0
fastapi==0.128.0
uvicorn==0.40.0
supabase==2.27.2
python-dotenv==1.2.1
python-multipart==0.0.22
streamlit==1.39.0
pandas==2.2.2
plotly==5.24.0
google-genai==0.8.3
markdown==3.7
weasyprint==62.3
